requirements.txt
python-dotenv
nextcord